Add the core interface and the TilePlanar codec
The core interface includes the BitDepth type and the Codec and
SettableImage interfaces. This defines the basic structure of how to set
up which format you need, such that later code can build on it.

The TilePlanar codec implements a planar image format that is used by
some well-known early consoles for their tile and sprite graphics.

This also includes some tests for the concrete types.