Add the TileRowPairPlanar codec
This encoding is used by at least one well-known retro console for some
of its tile graphics formats.

The name of this codec is not very clear, but I have not been able to
come up with a better name. Mainly because the format itself is a bit
tricky, being a mix of the per-row and per-tile planar formats.