Simplify the tests for Packed.Encode
This makes the tests be much easier to follow (less complicated), and
makes them use known data instead of relying on a second possibly-buggy
implementation. It also sets up for other codecs reusing the same code.

They are now less exhaustive, in that they only test from two locations
of the source image, and only use one set of pixels. However, seeing as
the code coverage is still 100%, that should be fine.