Simplify the tests for Packed.Decode
Much like for Encode, this reduces the number of locations that are
tested, along with only using one set of data (per bit depth), but in
return it adds testing source slices of different length and capacity.

This also improves the errors for corrupted source images in Encode, by
reusing the code this adds to verify the Decode output images.