Improve tests for Packed.Encode and Decode
The Encode tests now include testing for the tail of the destination
buffer being corrupted (similarly as for the Decode source), and the
Decode tests now report separately whether the source corruption was in
the source data or in the tail of trailing data after it.

This adds another image location to the Encode tests, since another test
was needed anyway to fully verify the buffer corruption. Thus, both the
Encode and Decode tests now check 3 image locations per bit depth.