Simplify tests for TilePlanar.Encode and Decode
This uses the same framework as the Packed tests, with different data.

This also adds a check into the framework functions to verify that the
passed data is the correct length for the codec, since their tests
actually rely on that for some of their correctness. It is verified
against Codec.Size(), so if that is wrong, it must be fixed first.