Simplify the Encode tests
This makes it use pieces of the new framework, replacing the old one.

As a result, it now runs fewer tests, only testing with one set of data
instead of two, but I think that's plenty for what this test does.
(And if not, it wouldn't be difficult to add a test with other data.)

This also removes the old testImageData type, which is no longer used.
It was overly complicated (essentially reimplementing the codecs) for
what it was supposed to do, and while the new tests are less exhaustive
in some ways, in other ways they actually do more, in much less time.